% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_a_regex_brick.R
\name{build_a_regex_brick}
\alias{build_a_regex_brick}
\title{Build a regex brick}
\usage{
build_a_regex_brick(
  type = c("ordered", "unordered"),
  content = c("wildcard", "lowercase letters", "uppercase letters", "digits",
    "punctuation", "space and tab", "custom"),
  occurrence = c("once", "at least once", "anytime", "custom"),
  custom_motif = NULL,
  custom_occurrence = NULL,
  escape = TRUE
)
}
\arguments{
\item{type}{character. The type of regex group.
Can be "ordered" or "unordered".}

\item{content}{character. The content of the regex group.
Can be "wildcard", "lowercase letters", "uppercase letters", "digits",
"punctuation", "space and tab", or "custom".}

\item{occurrence}{character. The occurrence of the regex group.
Can be "once", "at least once", "anytime", or "custom".}

\item{custom_motif}{character. The custom motif to be used in a regex group.
Only applicable if "custom" is selected for the content parameter.}

\item{custom_occurrence}{integer. The occurrence to be used in a regex group.
Only applicable if "custom" is selected for the occurrence parameter.}

\item{escape}{logical. Should special characters in custom motif be escaped.}
}
\value{
A character string representing the built regex.
}
\description{
Build a regex brick
}
\examples{
build_a_regex_brick(
	type = "ordered",
	occurrence = "custom",
	custom_motif = "hello$",
	custom_occurrence = c(1, 5),
	escape = FALSE
)

build_a_regex_brick(
	type = "unordered",
	occurrence = "anytime",
	custom_motif = "hello!",
	content = c("punctuation", "custom")
)
}
