% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_color_pal.R
\docType{data}
\name{color_pal}
\alias{color_pal}
\title{color_pal}
\format{
A data frame with  rows and  variables:
\describe{
}
}
\source{
Source
}
\usage{
color_pal
}
\description{
Character vector of 20 gold/purple colors
}
\keyword{datasets}
