% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_from_brick.R
\name{generate_from_brick}
\alias{generate_from_brick}
\title{Generate a word from a regex brick}
\usage{
generate_from_brick(brick, use_scrabble = TRUE, maxword = 10, maxfreq = 10)
}
\arguments{
\item{brick}{list. A list of the regex specifics}

\item{use_scrabble}{logical. Should the regex try to match a real word}

\item{maxword}{integer. Maximum number of random word to return}

\item{maxfreq}{integer. Boundary when no maximum occurrence is specified}
}
\value{
wordlist. A character string matching the regex.
}
\description{
Generate a word from a regex brick
}
\examples{
brick <- build_a_regex_brick(
  type = "ordered",
  occurrence = "custom",
  custom_motif = "heyho",
  content = NULL,
  custom_occurrence = c(5, 10)
)

generate_from_brick(brick)

brick <- build_a_regex_brick(
  type = "unordered",
  occurrence = "at least once",
  custom_motif = "write your motif here",
  content = c("digit")
)

generate_from_brick(brick)

brick <- build_a_regex_brick(
  type = "unordered",
  occurrence = "anytime",
  custom_motif = "WonDERFul",
  content = c("custom")
)

generate_from_brick(brick)
}
