% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lurex-package.R
\docType{package}
\name{lurex-package}
\alias{lurex}
\alias{lurex-package}
\title{lurex: Learn to Use Regular EXpression}
\description{
A simple app to help build simple regex.
}
\author{
\strong{Maintainer}: Swann Floc'hlay \email{swann@thinkr.fr}

}
\keyword{internal}
