% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_regex.R
\name{validate_regex}
\alias{validate_regex}
\title{Validate a regex combination}
\usage{
validate_regex(bricks)
}
\arguments{
\item{bricks}{character. The ordered list of regex bricks}
}
\value{
logical. Side effect : trigger a toast widget
}
\description{
Validate a regex combination
}
\examples{
validate_regex(bricks = list("yes", "or", "no"))
}
