% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_wordlist.R
\docType{data}
\name{wordlist}
\alias{wordlist}
\title{wordlist}
\format{
A character vector of lower case english words
\describe{
}
}
\source{
http://norvig.com/ngrams/
}
\usage{
wordlist
}
\description{
Word list.
}
\keyword{datasets}
