% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_regex.R
\name{validate_regex}
\alias{validate_regex}
\title{Validate a regex combination}
\usage{
validate_regex(bricks, use_toast = TRUE)
}
\arguments{
\item{bricks}{character. The ordered list of regex bricks}

\item{use_toast}{logical. Should a UI toast be shown for incorrect regex.}
}
\value{
list. logical indicating validity, character indicating reasons
}
\description{
Validate a regex combination
}
\examples{
validate_regex(
  bricks = list("1", "end", "or", "2", "end", "or", "3", "4", "end")
)
}
